Add date and human readable numbers support
Problem: A majority of time graphs are produced with dates on the x-axis and numbers on the y-axis. By default gnuplot does not interpret dates as datetimes, but tries to parse them as numbers. This causes the graphs to look wonky because it only interprets the year.

Problem: When dealing with numbers the default format for y-axis labels is scientific notation when numbers go beyond a certain scale. This is not easily readable by the average consumer of these graphs.

Fix: Add a flag to denote the x-axis is a datetime column.

Fix: Add a flag to format the y-axis labels in a human readable format (e.g. 100, 10k, 10M, 10B, etc).